lock down users mutations

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -56,7 +56,8 @@ const Mutation = {
   },
 
   async deleteUser(parent, args, {
-    prisma
+    prisma,
+    request
   }, info) {
     // const userExists = await prisma.exists.User({
     //   id: args.id
@@ -65,10 +66,11 @@ const Mutation = {
     // if (!userExists) throw new Error("User not found");
 
     // code above not necessary, left for reference sake
+    const userId = getUserId(request);
 
     return await prisma.mutation.deleteUser({
         where: {
-          id: args.id
+          id: userId
         }
       },
       info
@@ -76,11 +78,14 @@ const Mutation = {
   },
 
   updateUser(parent, args, {
-    prisma
+    prisma,
+    request
   }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.updateUser({
         where: {
-          id: args.id
+          id: userId
         },
         data: args.data
       },
@@ -112,8 +117,20 @@ const Mutation = {
   },
 
   async deletePost(parent, args, {
-    prisma
+    prisma,
+    request
   }, info) {
+    const userId = getUserId(request);
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) throw new Error("Unable to delete post");
+
     return await prisma.mutation.deletePost({
         where: {
           id: args.id
